Create FlowingParticles
FlowParticles is a python script that runs on the iphone App Pythonista. When the screen is touched, colorful bubble like particles appear, grow in size, change color, change transparency, and flow to the upper left corner of the screen with additional vector magnitude away form each other if many particles are formed in the same spot. Built to look pretty and entertain a small child.